refactor(opportunity): рефактор компонента в целях повышения универсальности
