+++ src/components/donation-page/opportunities/index.ts
@@ -1,0 +1,1 @@
+export * from './opportunities';
