+++ src/components/donation-page/opportunities/opportunities.module.css
@@ -1,0 +1,5 @@
+.opportunities {
+  padding: scale(173px) 0 0 0;
+  border: 2px solid green;
+  margin: 0;
+}
