--- src/components/donation-page/opportunity/index.ts
@@ -1,1 +1,0 @@
-export * from './opportunity';
