+++ src/pages/donation/index.tsx
@@ -3,7 +3,7 @@ import Head from 'next/head';
 
 import AppLayout from 'components/app-layout';
 import { Enthusiasm } from 'components/donation-page/enthusiasm';
-import { Opportunity } from 'components/donation-page/opportunity';
+import { Opportunities } from 'components/donation-page/opportunities';
 import { Report } from 'components/donation-page/report';
 
 import mockData from './assets/mock-donation-data.json';
@@ -19,15 +19,8 @@ const Donation: NextPage = () => (
         text={mockData.enthusiasm.text}
         footnote={mockData.enthusiasm.footnote}
       />
-      <Opportunity
-        title={mockData.opportunities[0].title}
-        picture={mockData.opportunities[0].picture}
-        kickies={mockData.opportunities[0].kickies}
-      />
-      <Opportunity
-        title={mockData.opportunities[1].title}
-        picture={mockData.opportunities[1].picture}
-        kickies={mockData.opportunities[1].kickies}
+      <Opportunities
+        data={mockData.opportunities}
       />
       <Report
         text={mockData.report.text}
